Symphony Fluid 960 Grid System Ensemble: initialize repository
This initial commit of the ensemble was created from an expanded ZIP archive
generated by the Export Ensemble. It allows switching between JavaScript libraries:
MooTools 1.2.1 and jQuery 1.3.1. Note: the MooTools library is not compressed.

This ensemble will be used to create updated versions of the grid layouts
for designinfluences.com